requirements.txt
# Core CV & Detection
opencv-python==4.9.0.80
ultralytics==8.2.18
Pillow==10.3.0

# gRPC
grpcio==1.64.0
grpcio-tools==1.64.0
protobuf==4.25.3

# Vector Store & Embeddings
faiss-cpu==1.8.0
sentence-transformers==3.0.1
numpy==1.26.4

# LLM
anthropic==0.28.0

# API
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
websockets==12.0

# Utilities
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.3.0
loguru==0.7.2
rich==13.7.1

# Testing
pytest==8.2.0
pytest-asyncio==0.23.7
httpx==0.27.0
